functions/requirements.txt
firebase_functions~=0.1.0
youtube-transcript-api
firebase-admin
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
requests
pytz